add react router and add some links for multiple pages
