+++ web/src/App.jsx
@@ -1,62 +1,28 @@
-import { useState } from 'react';
-import { TaskList } from './components/domains/task/TaskList/TaskList';
-import { NewTaskForm } from './components/domains/task/TaskItem/NewFormTask';
+import { Link, Outlet } from 'react-router';
+import './App.css';
+import logo from './logo.svg';
 
 function App() {
-  const [taskList, setTaskList] = useState([
-    {
-      id: 1,
-      title: 'Re-work UI/UX',
-      priority: 'low',
-      dueDate: '12/04/2021',
-      members: ['Said', 'Rechal'],
-      description: 'Timmer App',
-    },
-    {
-      id: 2,
-      title: 'Dark Mode Toggle',
-      priority: 'low',
-      dueDate: '12/05/2025',
-      members: ['Umair', 'Precious'],
-      description: 'Asa Dark Mode Feature',
-    },
-    {
-      id: 3,
-      title: 'Assessibility Checks',
-      priority: 'high',
-      dueDate: '11/03/2025',
-      members: ['Ebtesam', 'Deborah'],
-      description: 'Timer App',
-    },
-    {
-      id: 4,
-      title: 'Notification Checks',
-      priority: 'medium',
-      dueDate: '15/04/2025',
-      members: ['Michel', 'Ricardo'],
-      description: 'Timer App',
-    },
-  ]);
-  const newTask = (taskName, projectName, event) => {
-    event.preventDefault();
-    setTaskList((prevTaskInfo) => [
-      ...prevTaskInfo,
-      {
-        id: `${prevTaskInfo.length + 1}`,
-        title: taskName,
-        priority: 'High',
-        dueDate: '11/03/2025',
-        members: ['Syed', 'Said'],
-        description: projectName,
-      },
-    ]);
-  };
   return (
-    <>
-      <TaskList tasks={taskList} />
-      <NewTaskForm handelButton={newTask} />
-    </>
+    <div className="container">
+      {/* Sidebar */}
+      <div className="sidebar">
+        <img src={logo} alt="Logo" className="logo" />
+        <ul className="nav-links">
+          <li>
+            <Link to="/task-manager">Task Manager</Link>
+          </li>
+          <li>
+            <Link to="/add-task">Add to Task</Link>
+          </li>
+        </ul>
+      </div>
+
+      {/* Content Area */}
+      <div className="content">
+        <Outlet />
+      </div>
+    </div>
   );
 }
-
 export default App;
