+++ web/src/main.jsx
@@ -4,14 +4,17 @@ import { BrowserRouter, Route, Routes } from 'react-router';
 import App from './App.jsx';
 
 import './index.css';
-import { TempPage } from './pages/TempPage.jsx';
+import { TaskManager } from './pages/TaskManager.jsx';
+import { AddToTask } from './pages/AddToTask.jsx';
 
 createRoot(document.getElementById('root')).render(
   <StrictMode>
     <BrowserRouter>
       <Routes>
-        <Route path="/" element={<App />} />
-        <Route path="/temppage" element={<TempPage />} />
+        <Route path="/" element={<App />}>
+          <Route path="/task-manager" element={<TaskManager />} />
+          <Route path="/add-task" element={<AddToTask />} />
+        </Route>
       </Routes>
     </BrowserRouter>
   </StrictMode>,
