--- web/src/pages/TempPage.jsx
@@ -1,7 +1,0 @@
-export const TempPage = () => {
-  return (
-    <div>
-      <h1>TempPage</h1>
-    </div>
-  );
-};
